Remove unnecessary events during ssl_connect() phase
Keeping an extra event in the polling list can make the socket ready
immediately. This leads to 100% CPU consumption because the polling
function won't wait.

This is exactly the case for a large number of TLS connections. It can
take minutes until all the TLS sessions are established and we don't
want to consume all CPU cores in vain.