Complete WARMUP phase when all threads are ready
When the 1st thread transits to NORMAL phase it starts generating
intensive traffic that affects TLS handshakes in other threads.
Therefore, when we suppose to finish all TLS handshakes in ~5min, some
threads cannot finish in 10min.

To resolve this, synchronize between threads and transit to NORMAL phase
in all threads about simultaneously.